Adds ESP32-S2, ESP32-S3, ESP32-C3 and ESP32-H2 as being supported MCSs
Instead of reinventing the wheel with the ADC channels I decided to use the mechanics that micropython has in place. It works and it wasn't hard to update the code to use those mechanics.

Once past that I ran into issues with macro names not being available. This was easily handled as there were macros that did exist to do the same tasks and they just had different names.

The last step which was a little more of a challenge was the linking errors. This was taken care of in a couple of different ways. I used a combination of defining some of the missing functions, setting rules so gen_mpy wouldn't process some of the header files and the last thing was I write a script that removes specific functions from the lv_idf.c file. The script is automatically run so no additional things need to be done.

There is no special handling that needs to take place for the new modules. In fact the ESP32 is no longer limited to only using GPIOs that are attached to ADC1. any ADC pin is able to beused now. The ADC pins are used for the touch interfaces.

I have not updated the documentation yet as I would like someone to test and see if it works. I do not have an SPI screen to test with.